Revert "fix: add maintenance message for swap and explorer (#7327)"
This reverts commit c4c1a4a1fe2bc30a5c6abbd0144c972924f25a7d.
