+++ apps/cowswap-frontend/index.html
@@ -1,53 +1,97 @@
-<!DOCTYPE html>
+<!doctype html>
+
 <html lang="en">
+  <head>
+    <title>CoW Swap | The smartest way to trade cryptocurrencies</title>
+    <meta charset="utf-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1" />
+    <meta
+      name="description"
+      content="CoW Swap finds the lowest prices from all decentralized exchanges and DEX aggregators & saves you more with p2p trading and protection from MEV"
+    />
+    <script src="/emergency.js"></script>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>cow.fi Maintenance</title>
-  <style>
-    body {
-      margin: 0;
-      min-height: 100vh;
-      display: grid;
-      place-items: center;
-      font-family: Arial, sans-serif;
-      background: #111;
-      color: #fff;
-    }
+    <!-- Manifest, theme & colors  -->
+    <link rel="apple-touch-icon" sizes="180x180" href="/apple-touch-icon.png" />
+    <link rel="icon" type="image/png" href="/favicon-light-mode.png" media="(prefers-color-scheme: light)" />
+    <link rel="icon" type="image/png" href="/favicon-dark-mode.png" media="(prefers-color-scheme: dark)" />
+    <link rel="icon" type="image/png" href="/favicon-light-mode.png" media="(prefers-color-scheme: no-preference)" />
 
-    .notice {
-      max-width: 720px;
-      padding: 32px;
-      margin: 24px;
-      border: 1px solid #333;
-      border-radius: 12px;
-      background: #1a1a1a;
-      text-align: center;
-      box-shadow: 0 8px 30px rgba(0, 0, 0, 0.35);
-    }
+    <link rel="manifest" href="/manifest.json" />
+    <link rel="mask-icon" href="/safari-pinned-tab.svg" color="#000000" />
+    <meta name="msapplication-TileColor" content="#000000" />
+    <meta name="theme-color" content="#ffffff" />
 
-    h1 {
-      margin: 0 0 16px;
-      font-size: 2rem;
-      line-height: 1.2;
-    }
+    <!-- Preload critical fonts -->
+    <link rel="preload" href="/static/Inter-roman.var.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/Inter-italic.var.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/StudioFeixenSans-Regular.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/StudioFeixenSans-Medium.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/StudioFeixenSans-Semibold.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/StudioFeixenSans-Bold.woff2" as="font" type="font/woff2" crossorigin />
+    <link rel="preload" href="/static/StudioFeixenMono-Regular.woff2" as="font" type="font/woff2" crossorigin />
 
-    p {
-      margin: 0;
-      font-size: 1.125rem;
-      line-height: 1.6;
-    }
-  </style>
-</head>
+    <!-- Social media tags -->
+    <meta property="og:type" content="website" />
+    <meta property="og:title" content="CoW Swap | The smartest way to trade cryptocurrencies" />
+    <meta
+      property="og:description"
+      content="CoW Swap finds the lowest prices from all decentralized exchanges and DEX aggregators & saves you more with p2p trading and protection from MEV"
+    />
+    <meta property="og:image" content="https://swap.cow.fi/images/og-meta-cowswap.png?v=4" />
+    <meta property="og:url" content="https://swap.cow.fi" />
 
-<body>
-  <main class="notice">
-    <h1>Maintenance</h1>
-    <p>
-      The cow.fi domain is currently under maintenance. We'll be back shortly.
-    </p>
-  </main>
-</body>
+    <!-- Twitter media tags -->
+    <meta name="twitter:card" content="summary_large_image" />
+    <meta name="twitter:site" content="@CoWSwap" />
+    <meta name="twitter:title" content="CoW Swap | The smartest way to trade cryptocurrencies" />
+    <meta name="twitter:image" content="https://swap.cow.fi/images/og-meta-cowswap.png?v=4" />
+    <style>
+      html,
+      body {
+        margin: 0;
+        padding: 0;
+      }
+    </style>
+  </head>
 
+  <body>
+    <noscript>
+      <style>
+        main {
+          margin: 0 auto;
+        }
+
+        main img {
+          width: 100%;
+        }
+
+        .banner {
+          padding: 10px;
+          background-color: #fdfda8;
+          border-radius: 16px;
+        }
+      </style>
+      <main>
+        <img src="https://swap.cow.fi/images/og-meta-cowswap.png?v=4" alt="CoW Swap Logo" />
+        <p class="banner">
+          CoW Swap uses JavaScript. Please
+          <a href="https://support.google.com/adsense/answer/12654?hl=en" target="_blank" rel="noopener noreferrer"
+            >enable it in your browser</a
+          >
+          🐮
+        </p>
+        <ul>
+          <li><a href="https://swap.cow.fi/">Home | CoW Swap</a></li>
+          <li><a href="https://discord.com/invite/cowprotocol">Chat</a></li>
+          <li><a href="https://docs.cow.fi">Docs</a></li>
+          <li><a href="https://dune.com/cowprotocol/cowswap">Stats</a></li>
+          <li><a href="https://twitter.com/CoWSwap">Twitter (@CoWSwap)</a></li>
+        </ul>
+      </main>
+    </noscript>
+
+    <div id="root"></div>
+    <script type="module" src="/src/main.tsx"></script>
+  </body>
 </html>
